Fix forward command getting stuck when capital letter has non-letters and the bol before it.
Found that point got stuck with a sentence that had garbage characters.
E.g. with | representing the point:
This is a |sentence.
"..Rest of sentence"

The point would move backwards to "This" because
sentence-nav--maybe-before-sentence-start-p would return nil until then.
For a line to be considered a comment, require a space afterwards to prevent this.